Adding of code seperation
